requirements.txt
customtkinter
tkcalendar
pillow
sqlite3